moved the get_data_type and get_path_leaf to models.py from views.py to clean up
